slack_sensor/requirements.txt
rid-lib
python-dotenv
slack_bolt
fastapi
uvicorn